Issue #14 serde_utils-based (de)serialization of OpenClosed, OnOff
Prior to this patch, empty variants such as "OpenClosed::On" were
serialized as "On: []", which is quite surprising. This patch
overrides the default behavior of OpenClosed and OnOff to make sure
that the empty variants are serialized simply as strings.

For this purpose, the patch introduces a dependency to
`serde_utils`. If this proves troublesome, we may be able to
eventually get rid of this dependency.