Resolves #45 - Enable concurrency for to the AdapterManager.
The current version of the AdapterManager serializes all requests to a
single thread. In particular, any (synchronous) interaction with a
device blocks every (also synchronous) access to the
AdapterManager. This is not very good for performance.

This patch refactors the AdapterManager to enable concurrent use.

The AdapterManager is now Sync + Send and can be accessed from any
thread. The state of the AdapterManager is protected by a RwLock-style
lock that lets any number of concurrent clients access the read-only
state, or a single client access the read/write state. Furthermore,
all access to devices takes place outside of the locks, hence ensuring
that it does not block other clients.

The case of watch is more complex, as the AdapterManager needs to
update its state in two further circumstances:

- after placing a call to the Adapters (and possibly devices);
- whenever a WatchGuard is dropped.

These two kinds of updates can race with each other. For this reason,
watch-related updates are serialized to a dedicated background thread,
which handles both the call to the Adapters, the updates performed
after placing the call and the updates performed when a WatchGuard is
dropped.

This patch does not implement issue #47.